prac5

+++ 240422_project/prac4.cpp
@@ -1,9 +1,12 @@
+//240422
+// ctrl + c + K : 전체 주석, ctrl + k + u : 전체주석해제
 #include <iostream>
 
 int main()
 {
+	
 	std::cout << "첫 번째 변수를 입력하세요.";
-	int x;
+	float x;
 	std::cin >> x;
 
 	std::cout << "원하는 연산자를 고르세요. + : 0, - : 1, * : 2, / :3 \n";
@@ -11,7 +14,7 @@ int main()
 	std::cin >> opr;
 
 	std::cout << "두 번째 변수를 입력하세요.";
-	int y;
+	float y;
 	std::cin >> y;
 
 	switch (opr) {
@@ -27,8 +30,9 @@ int main()
 		case 3:
 			std::cout << x / y;
 			break;
-
 	}
+	
+
 
 
 }
